test: 🚨 improve template tests

+++ src/feature/__tests__/cleanForTests.ts
@@ -1,6 +1,7 @@
 import { searchFilesInDirectory } from '@/feature/__tests__/searchFilesInDirectory';
 import { createPath } from '@/util/createPath';
 import { deletePath } from '@/util/deletePath';
+import { isFolderExist } from '@/util/isFolderExist';
 
 export const cleanUpFiles = async ({
   snpCatalog, //TODO - make it reusable
@@ -11,9 +12,25 @@ export const cleanUpFiles = async ({
   directoryPath: string;
   isDebug: boolean;
 }): Promise<void> => {
-  await deletePath(createPath(snpCatalog), isDebug);
-  const allFiles = await searchFilesInDirectory({ directoryPath });
-  for (const file of allFiles) {
-    await deletePath(createPath(file), isDebug);
+  if (
+    await isFolderExist({
+      folderPath: snpCatalog,
+      createFolder: false,
+    })
+  ) {
+    await deletePath(createPath(snpCatalog), isDebug);
+  }
+
+  if (
+    await isFolderExist({
+      folderPath: directoryPath,
+      createFolder: false,
+    })
+  ) {
+    const allFiles = await searchFilesInDirectory({ directoryPath });
+    for (const file of allFiles) {
+      await deletePath(createPath(file), isDebug);
+    }
+    await deletePath(createPath(directoryPath), isDebug);
   }
 };
